+++ UnderAutomation.Fanuc.ObfuscatedSources/Snpx/Assignment/NumericRegistersBatchAssignment.cs
@@ -1,0 +1,28 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+using Snpx.Internal;
+using System;
+
+namespace Snpx.Assignment {
+
+	public class NumericRegistersBatchAssignment : BatchAssignment<float, int> {
+
+
+		public NumericRegistersBatchAssignment()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+		/// <summary>
+		/// Read all numeric registers assigned in this batch assignment.
+		/// </summary>
+		/// <returns>Returns an array of float values representing the numeric registers</returns>
+		public override float[] Read()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+			return default;
+		}
+	}
+}