+++ UnderAutomation.Fanuc.ObfuscatedSources/Snpx/Assignment/PositionSystemVariablesBatchAssignment.cs
@@ -1,0 +1,26 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+using Snpx.Internal;
+using Common;
+using System;
+
+namespace Snpx.Assignment {
+
+	public class PositionSystemVariablesBatchAssignment : BatchAssignment<Position, string> {
+
+
+		public PositionSystemVariablesBatchAssignment()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+
+		public override Position[] Read()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+			return default;
+		}
+	}
+}