+++ UnderAutomation.Fanuc.ObfuscatedSources/Snpx/Assignment/RealSystemVariablesBatchAssignment.cs
@@ -1,0 +1,25 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+using Snpx.Internal;
+using System;
+
+namespace Snpx.Assignment {
+
+	public class RealSystemVariablesBatchAssignment : BatchAssignment<float, string> {
+
+
+		public RealSystemVariablesBatchAssignment()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+
+		public override float[] Read()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+			return default;
+		}
+	}
+}