+++ UnderAutomation.Fanuc.ObfuscatedSources/Snpx/Internal/NumericRegisters.cs
@@ -3,10 +3,11 @@
 // The internal code is hidden; to access it, you need to obtain a Source licence of the library.
 
 using System;
+using Snpx.Assignment;
 
 namespace Snpx.Internal {
 
-	public class NumericRegisters : SnpxWritableAssignableElements<float, int> {
+	public class NumericRegisters : SnpxWritableAssignableElements<float, int, NumericRegistersBatchAssignment> {
 
 
 		protected override string GetAssignmentName(int index)
@@ -33,6 +34,13 @@ namespace Snpx.Internal {
 		protected override void WriteInClient(int offset, float value)
 		{
 			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+
+		public NumericRegistersBatchAssignment CreateBatchAssignment(int startIndex, int count)
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+			return default;
 		}